+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "adsr-envelope": "^1.0.0",
     "body-parse": "^0.1.0",
+    "body-parser": "^1.18.3",
     "codemirror": "^5.38.0",
     "ejs": "^2.6.1",
     "lodash": "^4.17.10",
